streamlit_chatbot_app/requirements.txt
streamlit==1.36.0
pyyaml>=6.0.1
pyarrow>=14,<17
